Move ServiceResolvable to handlers module
