+++ aries_vcx/src/messages/connection/mod.rs
@@ -3,4 +3,3 @@ pub mod invite;
 pub mod problem_report;
 pub mod request;
 pub mod response;
-pub mod service;
