+++ aries_vcx/src/messages/connection/request.rs
@@ -1,5 +1,5 @@
 use crate::messages::a2a::{A2AMessage, MessageId};
-use crate::did_doc::did_doc::*;
+use crate::did_doc::model::*;
 use crate::did_doc::DidDoc;
 use crate::messages::thread::Thread;
